Add implementation draft to read tldr from GitHub
There is now a draft implementation to read tldr man pages
directly from GitHub.

The plans how the plugin should eventually work are now
documented in the README file. The thought is that there
is a default that points to Linux tldr man pages. This
can be changed with the '--file' option to read the data
remotely or locally from a local file.

WIP

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>