Fix README and remove unnecessary commands
The README was simplified and the example to install the
Snippy tool was removed. The original intention was to
highlight that the plugin and tool must be installed into
the same Python namespace (local or global) to find the
plugins. Then it was thought that the default README has
to be as simple and short as possible without any bloat.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>