Add logic to parse URLs where to read tldr man pages
There is now draft logic to support various different ways
to provide an URL, file path or file with the '--file' CLI
option.

The idea is to be flexible for user. Because of this, the
code supports URLs and file paths in various formats like:

  --file https://github.com/tldr-pages/tldr/tree/0.0.1/pages
  --file ../tldr/pages/
  --file .
  --file file:///full/path/to/tldr/pages
  --file file:///full/path/to/tldr/pages/alpine.md
  --file ../tldr/pages/linux/alpine.md

The folder must point to URI or path that contains the tldr
pages, single tldr page or one file under the tldr page.
There is no plans to support recursive search of tldr man
pages from any random path.

WIP

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>