Add max-line-length for Pylint
The maximum line length for Pylint is now 120 characters. This
is because the 'snippy_import_hook' method terms table is
manually formatted with one long line. The Black code format
keeps the code formatting aligned to default black line length.

The manual formatting of the table is made in order to render
it nicely in documentation. The tables in rst are not very
"easy" to get correctly aligned. The width of the table is
too narrow if the manual formatting is not done. So far there
has been no better way figured out how to do this.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>