Fix Read the Docs auto-documentation
The snippy-tldr package is not installed in the Read the
Docs build and the modules cannot be found. The path where
the modules are located must be set with the sys.path [1].

It is an open question why this works without the sys.path
in the Snippy project. Is this automatically done by the
'sphinxcontrib.openapi' package in the Snippy project?

[1] https://stackoverflow.com/a/55019678
[2] http://www.sphinx-doc.org/en/master/usage/extensions/autodoc.html

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>