Refactor tldr page reading from GitHub
The tldr page parsing terms are now corrected according to
the tldr project. The term 'platform' is the one that is
officially reserved for platforms like 'linux' or 'windows'.
A 'page' refers to a single tldr page.

The parser does not rely anymore to HTML parsing. There is
a better way through the GitHub API. The API has a problem
that there is a throttling in GitHub that changes whether
the query is authenticated or not [1]. The implementation
tries to avoid using the GitHub API queries too much so
that the unauthenticated queries would be possible.

[1] https://developer.github.com/v3/#rate-limiting

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>