Fix test to verify parsing of default tldr files
There is now first test case that the verifies also the
parsing of tldr pages. The test verifies the basic usage
when the 'file' parameter is not provided for the plugin.

Refactored and cleaned the code. For example not used code
was removed, comments and code documentations were cleaned
and some variable names were changed.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>