Fix targets with pytests to use explicit folder
The Pytests commands now have explicit target where the tests
are located. If for some reason, there are other folder that
are accidentally cloned into same folder as the snippy-tldr
and they contain tests, the previous commands tried to run
all the tests it could found from the snippy-tldr root. Now
the Pytest commands are explicitly pointing to the snippy-
tldr project tests folder.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>