Fix Travis CI tests to install Snippy dependency
There is a fault in Snippy that requires installing also the
server dependencies when the new snippy.plugins is used. The
intention was that user could just install the 'snippy' to use
the plugins. But now the 'jsonschema' third party module in
the snippy.plugins is not installed by default with the snippy.
This dependency is installed currently only with the Snippy
server dependencies.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>