Refactor tests to avoid copying all HTTP requests
The tests includes several GitHub API requests and responses.
Defining all the HTTP requests and responses in every tests
adds line length for the tests.

Now the tests should be a bit more compact. This is always
a trade off on some level between different aspects. On the
other hand it is nice that tests include all the values
visible. But on the other hand, very long tests are a bit
hard to read. Also refactoring code with copy paste test
data may be hard if there are many tests.

As of now, a bit more compact tests is the way to go.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>