Edit TODO list to include regexp failure
There is a regexp failure to read local tldr page from the
local filesystem. The regexp matches but it gives 'None'
for platform when reading '/root/path/linux/adduser.md'.

Why the group is None? Why the regexp matches? Why this
works in https://pythex.org/?

The regexp is the RE_CATCH_LOCAL_TLDR_PAGES.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>