Fix Snippy install for the Snippy-tldr plugin tests
The Snippy with finalized plugins module is not yet released
into PyPI. Because of this, the Snippy is installed from the
master branch as of now. This allows installing the required
module to make the tests work.

Also the jsonschema package is installed in the snippy-tldr
plugin. This is not installed by the Snippy as of now when
the CLI version is used. But the snippy.plugins module need
this package. This likely should be in the Snippy tool. But
as of now, it is included in the Snippy-tldr. The main reason
is to reduce the amount of dependencies in Snippy. If user
wants to have the Snippy-tldr, the required dependency can
be installed with the plugin.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>