Fix test mocks for classes in snippy.plugins
The Snippy installation was not properly made. The Snippy
was installed from version 0.10.0 which did not have the
plugin related module. Now the installation is made
temporarily from the Snippy master branch and the related
module is now available.

Removed unnecessary mocks for Const class and imports from
the tests.conftest for the mocks.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>