Fix Travis CI configuration for Snippy installation
The Snippy is now installed by the Snippy-tldr setup.py and
there is no need to manually install the Snippy anymore in
the Travis CI script.

Add also test stages for MacOS and Windows and linting for
Python2.

Python3.8 is not the release package not the 'dev' release
in the Travis CI configuration.

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>