Fix install-devel make target on local environment
The Python 3.4 installation was failing locally with error
related to 'BadZipfile: File is not a zip file'. The reason
for this is unknown. This should work because the Snippy
is able to install with the Python 3.4 and most of the Python
modules are the same.

This was not failing in Travis CI with Python 3.4.

This workaround is now forced from the Makefile to allow
exactly the same installation command to work with all
the Python versions. [1]

Add Makefile parameter to define PIP proxy, cache and
upgrade parameters.

The --upgrade-strategy option is now set by default to the
'eager'. There is a problem with the Snippy being installed
from the Git. If there is already for example an version
'0.10.0' from Snippy, the --upgrade option does not enforce
upgrade to the version from the Git master branch. [2] [3]

[1] https://stackoverflow.com/a/41916324
[2] https://pip.pypa.io/en/stable/reference/pip_install/
[3] https://github.com/pypa/pip/issues/6457#issuecomment-520373857

Signed-off-by: Heikki Laaksonen <laaksonen.heikki.j@gmail.com>